#267. Refactoring.
