+++ common/src/commonMain/kotlin/io/xorum/codeforceswatcher/features/verification/redux/VerificationRequests.kt
@@ -1,7 +1,7 @@
-package io.xorum.codeforceswatcher.features.verification
+package io.xorum.codeforceswatcher.features.verification.redux
 
-import io.xorum.codeforceswatcher.features.auth.UserAccount
-import io.xorum.codeforceswatcher.network.VerificationRepository
+import io.xorum.codeforceswatcher.features.auth.models.UserAccount
+import io.xorum.codeforceswatcher.features.verification.VerificationRepository
 import io.xorum.codeforceswatcher.network.responses.backend.Response
 import io.xorum.codeforceswatcher.redux.*
 import tw.geothings.rekotlin.Action
