+++ common/src/commonMain/kotlin/io/xorum/codeforceswatcher/network/BackendRepository.kt
@@ -1,12 +1,9 @@
 package io.xorum.codeforceswatcher.network
 
 import io.ktor.client.request.parameter
-import io.xorum.codeforceswatcher.features.auth.UserAccount
 import io.xorum.codeforceswatcher.features.users.models.User
-import io.xorum.codeforceswatcher.features.verification.VerificationCodeResponse
 import io.xorum.codeforceswatcher.network.responses.backend.KtorResponseClient
 import io.xorum.codeforceswatcher.network.responses.backend.NewsResponse
-import io.xorum.codeforceswatcher.network.responses.backend.UserData
 import io.xorum.codeforceswatcher.redux.store
 
 internal class BackendRepository {
@@ -20,9 +17,9 @@ internal class BackendRepository {
                 parameter("version", "v2")
             }
 
-    suspend fun fetchUsers(handles: String, isAllRatingChangesNeeded: Boolean) =
+    suspend fun fetchUser(handle: String, isAllRatingChangesNeeded: Boolean) =
             ktorResponseClient.get<List<User>>("users") {
-                parameter("handles", handles)
+                parameter("handles", handle)
                 parameter("isAllRatingChangesNeeded", isAllRatingChangesNeeded.toString())
             }
