+++ common/src/commonMain/kotlin/io/xorum/codeforceswatcher/network/responses/backend/UserData.kt
@@ -1,6 +1,6 @@
 package io.xorum.codeforceswatcher.network.responses.backend
 
-import io.xorum.codeforceswatcher.features.auth.UserAccount
+import io.xorum.codeforceswatcher.features.auth.models.UserAccount
 import io.xorum.codeforceswatcher.features.users.models.User
 import kotlinx.serialization.Serializable
