+++ common/src/commonMain/kotlin/io/xorum/codeforceswatcher/redux/reducers/AppReducer.kt
@@ -5,7 +5,7 @@ import io.xorum.codeforceswatcher.features.news.redux.reducers.newsReducer
 import io.xorum.codeforceswatcher.features.users.redux.reducers.usersReducer
 import io.xorum.codeforceswatcher.features.problems.redux.reducers.problemsReducer
 import io.xorum.codeforceswatcher.features.contests.redux.reducers.contestsReducer
-import io.xorum.codeforceswatcher.features.verification.verificationReducer
+import io.xorum.codeforceswatcher.features.verification.redux.verificationReducer
 import io.xorum.codeforceswatcher.redux.states.AppState
 import tw.geothings.rekotlin.Action
